move jdk back to 1.7 update plugins and dependencies update gradle to 3.2
