+++ build.gradle
@@ -9,7 +9,7 @@ plugins {
     id 'idea'
 
     // intershop test plugin
-    id 'com.intershop.gradle.test' version '2.1.0'
+    id 'com.intershop.gradle.test' version '2.2.0'
 
     // plugin for documentation
     id 'org.asciidoctor.convert' version '1.5.3'
@@ -18,13 +18,13 @@ plugins {
     id 'maven-publish'
 
     // intershop version plugin
-    id 'com.intershop.gradle.scmversion' version '2.1.0'
+    id 'com.intershop.gradle.scmversion' version '2.3.0'
 
     // plugin for publishing to Gradle Portal
-    id 'com.gradle.plugin-publish' version '0.9.5'
+    id 'com.gradle.plugin-publish' version '0.9.6'
 
     // plugin for publishing to jcenter
-    id 'com.jfrog.bintray' version '1.7.1'
+    id 'com.jfrog.bintray' version '1.7.3'
 }
 
 scm {
@@ -39,8 +39,8 @@ version = scm.version.version
 
 // plugins should work with JDK 1.7 aslong
 // older INTERSHOP version use this version of JDK
-sourceCompatibility = 1.8
-targetCompatibility = 1.8
+sourceCompatibility = 1.7
+targetCompatibility = 1.7
 
 // set correct project status
 if (project.version.endsWith('-SNAPSHOT')) {
@@ -49,7 +49,7 @@ if (project.version.endsWith('-SNAPSHOT')) {
 
 test {
     // Gradle versions for test
-    systemProperty 'intershop.gradle.versions', '2.11,3.1'
+    systemProperty 'intershop.gradle.versions', '2.11,3.1,3.2'
 }
 
 task copyAsciiDoc {
@@ -221,11 +221,11 @@ pluginBundle {
 }
 
 dependencies {
-    compile 'com.intershop.gradle.nexusstaging:nexusstaging-gradle-plugin:2.1.0'
-    compile 'com.intershop.gradle.jiraconnector:jiraconnector-gradle-plugin:2.1.0'
-    compile 'com.intershop.gradle.repoconfig:repoconfig-gradle-plugin:1.1.1'
-    compile 'org.jfrog.buildinfo:build-info-extractor-gradle:4.4.7'
-    compile 'com.intershop.gradle.buildinfo:buildinfo-gradle-plugin:2.1.0'
+    compile 'com.intershop.gradle.nexusstaging:nexusstaging-gradle-plugin:2.2.0'
+    compile 'com.intershop.gradle.jiraconnector:jiraconnector-gradle-plugin:2.2.0'
+    compile 'com.intershop.gradle.repoconfig:repoconfig-gradle-plugin:3.1.0'
+    compile 'org.jfrog.buildinfo:build-info-extractor-gradle:4.4.9'
+    compile 'com.intershop.gradle.buildinfo:buildinfo-gradle-plugin:2.2.0'
 
     testCompile('org.spockframework:spock-core:1.0-groovy-2.4'){
         exclude group: 'org.codehaus.groovy'
